requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.25
yt-dlp>=2024.3.10
pyrogram>=2.0.106
tgcrypto>=1.2.5